refactor: split out basic api methods in BasicAPIClient
